Make the 1st space optional on fzf-insert-history
The old regex worked fine for me until I passed 9999 entries
in my history. Then it stopped matching because the first part
of the regex contained a mandatory space.

And that space was exactly where the 1 of 10000 appeared.

```
 9999* command 9999
10000* command 10000
^ look ma, no spaces!
```

So I basically corrected that by changing this

    's/ +[0-9]+\*?  ?//'
        ^
into this

    's/ *[0-9]+\*?  ?//'
        ^

By changing the `<space>+` into a `<space>*` in the
beginning, I made it optional to have a space in the beginning
of the entry. I also added a `^` in the beginning for good measure,
since it kind of gets a little more generic and even more
potentially-matching-unexpected-things than it was.